requirements.txt
torch>=1.8.0
open_clip
Pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.50.0
matplotlib>=3.3.0